requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
pillow==10.1.0
playwright==1.40.0
requests==2.31.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
